feat: home and about page, fix project card hover
