+++ apps/web/src/pages/About/About.css.ts
@@ -1,0 +1,8 @@
+import { style } from '@vanilla-extract/css';
+import { vars } from 'ui/variables.css';
+
+export const aboutStyles = style({
+  backgroundColor: vars.colors.bgPrimary,
+  height: '100%',
+  flex: '1',
+});
