+++ apps/web/src/pages/About/About.tsx
@@ -2,13 +2,13 @@ import { NextSeo } from 'next-seo';
 
 import Members from '@/components/Members';
 
-import { StyledAbout } from './styles';
+import { aboutStyles } from './About.css';
 
 const About = () => (
-  <StyledAbout>
+  <div className={aboutStyles}>
     <Members />
     <NextSeo title='About Project Shelf' />
-  </StyledAbout>
+  </div>
 );
 
 export default About;
