--- apps/web/src/pages/About/styles.ts
@@ -1,7 +1,0 @@
-import { styled } from '@/stitches/stitches.config';
-
-export const StyledAbout = styled('div', {
-  backgroundColor: '$bgPrimary',
-  height: '100%',
-  flex: '1',
-});
